Refactor activity validators
