+++ core/domain/prod_validation_jobs_one_off.py
@@ -19,8 +19,9 @@
 from __future__ import absolute_import  # pylint: disable=import-only-modules
 from __future__ import unicode_literals  # pylint: disable=import-only-modules
 
+import importlib
+
 from core import jobs
-from core.domain import prod_validators
 from core.platform import models
 import python_utils
 
@@ -92,7 +93,21 @@ class ProdValidationAuditOneOffJob( # pylint: disable=inherit-non-class
         if not model_instance.deleted:
             model_name = model_instance.__class__.__name__
             validator_cls_name = '%sValidator' % model_name
-            validator = getattr(prod_validators, validator_cls_name)
+            # Module name for models is of the form:
+            # 'core.storgae.<model-type>.gae_models'.
+            # Module name for validators is of the form:
+            # 'core.domain.<model-type>'.
+            # So, we extract the module name for models to obtain the module
+            # name for validators. There is no extra test required to verify
+            # that models and validators have names defined based on model-type
+            # since if they don't the validators test will automatically fail
+            # based on the import we perform here for validators.
+            model_module_name = model_instance.__module__
+            model_type = model_module_name.split('.')[2]
+            validator_module_name = '%s_validators' % model_type
+            validator_module = importlib.import_module(
+                'core.domain.%s' % validator_module_name)
+            validator = getattr(validator_module, validator_cls_name)
             validator.validate(model_instance)
             if len(validator.errors) > 0:
                 for error_key, error_val in validator.errors.items():
